// LR-4/react-frontend/src/api/Api.ts
export interface SmartOrder {
  /** @example 1 */
  id?: number;
  /** @example "formed" */
  status?: "draft" | "formed" | "completed" | "rejected" | "deleted";
  /** @example "ул. Примерная, д. 1, кв. 5" */
  address?: string;
  /**
   * @format float
   * @example 0.55
   */
  total_traffic?: number;
  /** @example 1 */
  client_id?: number;
  /** @example "client1" */
  client_name?: string;
  /**
   * @format date-time
   * @example "2025-10-21T13:13:40Z"
   */
  formed_at?: string;
  /**
   * @format date-time
   * @example "2025-10-21T13:14:20Z"
   */
  completed_at?: string;
  /** @example 2 */
  moderator_id?: number;
  /** @example "moderator1" */
  moderator_name?: string;
  /**
   * @format date-time
   * @example "2025-10-21T13:08:04Z"
   */
  created_at?: string;
  items?: OrderItem[];
}

export interface OrderItem {
  /** @example 2 */
  device_id?: number;
  /** @example "Лампочка" */
  device_name?: string;
  /** @example 3 */
  quantity?: number;
  /**
   * @format float
   * @example 0.5
   */
  data_per_hour?: number;
  /** @example "http://localhost:9000/image/lamp.png" */
  namespace_url?: string;
}